refactor sui code from gemstone
