+++ crates/gem_sui/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "gem_sui"
+version = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+
+bcs = "0.1.6"
+primitives = { path = "../primitives" }
+sui-types = { git = "https://github.com/mystenlabs/sui", package = "sui-types", tag = "mainnet-v1.27.4" }
+
+blake2 = { workspace = true }
+hex = { workspace = true }
+thiserror = { workspace = true }
+base64 = { workspace = true }
+anyhow = { workspace = true }
